+++ CHANGELOG.md
@@ -2,12 +2,11 @@
 
 ## [upcoming]
 
-- Inital decoder implementation
+- Decoder Inter support
 - Static library support
 - Alt-ref pictures - temporal filtering
 - Adaptive Transform Block for INTRA
 - Adaptive QP scaling
-- Decoder - Support for Tiles and 10bit
 - API - add option to calculate / report PSNR values
 - Support for segmentation